refactor: move modules tests
